0.1.2 - adds options V2 page, allows whitelisting
After using this extension for a number of years, I've run across a
handful of sites that use .gifs as button, etc. This extension wrongfully
targets them and often accidentally breaks the formatting and
functionality of the website.

My fix simply implements a whitelist of sites that disable the content
scripts through programmatic injection. Within background.js, my edits
check a parsed version of the contents of the whitelist versus the current
domain. If there are matches, the extension never executes any changes on
the page.

 - Contribution by Zach Hardesty: http://zachhardesty.com